+++ gwbackupy/tests/test_file_storage.py
@@ -1,3 +1,4 @@
+import datetime
 import io
 import os
 import tempfile
@@ -129,3 +130,94 @@ def test_link_remove():
         link_scanned = links.find(f=lambda link: link.id() == "testid")
         assert link_scanned is not None
         assert link_scanned.is_deleted()
+
+
+def test_find_not_valid_files():
+    with tempfile.TemporaryDirectory(prefix="myapp-") as temproot:
+        fs = FileStorage(root=temproot)
+        with open(os.path.join(temproot, ".gitignore"), "w") as f:
+            f.write("a")
+        with open(os.path.join(temproot, "without-extension"), "w") as f:
+            f.write("b")
+        links = fs.find()
+        assert len(links) == 0
+
+
+def test_find_with_tempfile():
+    with tempfile.TemporaryDirectory(prefix="myapp-") as temproot:
+        fs = FileStorage(root=temproot)
+
+        # tempfile remove test
+        tmp = os.path.join(temproot, "any.tmp")
+        with open(tmp, "w") as f:
+            f.write("a")
+        links = fs.find()
+        assert len(links) == 0
+        assert not exists(tmp)
+
+        # tempfile remove fail test
+        tmp2 = os.path.join(temproot, "any2.tmp")
+        with open(tmp2, "w") as f:
+            f.write("a")
+            links = fs.find()
+            assert len(links) == 0
+        assert exists(tmp2)
+
+
+def test_find_without_properties():
+    with tempfile.TemporaryDirectory(prefix="myapp-") as temproot:
+        fs = FileStorage(root=temproot)
+        with open(os.path.join(temproot, "testid.ext"), "w") as f:
+            f.write("a")
+        links = fs.find()
+        assert len(links) == 1
+        assert len(links[0].get_properties()) == 0
+
+
+def test_link_set_properties():
+    with tempfile.TemporaryDirectory(prefix="myapp-") as temproot:
+        fs = FileStorage(root=temproot)
+        link = fs.new_link("test", "ext")
+        link.set_properties({"my": "value"})
+        assert link.get_property("my") == "value"
+        fs.put(link, "data")
+        links = fs.find()
+        assert links[0].get_property("my") == "value"
+
+
+def test_eq():
+    with tempfile.TemporaryDirectory(prefix="myapp-") as temproot:
+        fs = FileStorage(root=temproot)
+        link = fs.new_link("test", "ext")
+        assert link != 4
+        link2 = fs.new_link("test2", "ext")
+        assert link != link2
+        time.sleep(0.002)
+        link3 = fs.new_link("test", "ext")
+        assert link != link3
+
+
+def test_subdirs():
+    with tempfile.TemporaryDirectory(prefix="myapp-") as temproot:
+        fs = FileStorage(root=temproot)
+        link = fs.new_link("test", "ext", datetime.datetime.now().timestamp())
+        assert os.path.dirname(link.get_file_path()) != temproot
+        assert os.path.dirname(link.get_file_path()).startswith(temproot)
+
+
+def test_remove_permanently_fail():
+    with tempfile.TemporaryDirectory(prefix="myapp-") as temproot:
+        fs = FileStorage(root=temproot)
+        link = fs.new_link("test", "ext")
+        fs.put(link, "data")
+        with open(link.get_file_path(), "w") as f:
+            assert not fs.remove(link, as_new_mutation=False)
+
+
+def test_remove_fail():
+    with tempfile.TemporaryDirectory(prefix="myapp-") as temproot:
+        fs = FileStorage(root=temproot)
+        link = fs.new_link("test", "ext")
+        fs.put(link, "data")
+        os.remove(link.get_file_path())
+        assert not fs.remove(link)