+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## DEV
 
+
+## 0.9.0
+
 - Bug: fix file storage new mutation as deleted with put fail error 
 - Enh #23: extend storage interface with modify
 - Enh #56: content hash added to gmail message objects